refactor(columnar): extract join operators
