+++ bench/meson.build
@@ -32,6 +32,7 @@ bench_backend_src = files(
   subdir_wirelog / 'columnar/arithmetic.c',
   subdir_wirelog / 'columnar/expr.c',
   subdir_wirelog / 'columnar/filter.c',
+  subdir_wirelog / 'columnar/join.c',
   subdir_wirelog / 'columnar/eval_stack.c',
   subdir_wirelog / 'columnar/eval.c',
   subdir_wirelog / 'columnar/arrangement.c',
